% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmicrobiologyfunction-package.R
\docType{package}
\name{bacterialgrowth-package}
\alias{bacterialgrowth}
\alias{bacterialgrowth-package}
\title{bacterialgrowth: bacterialgrowth Package}
\description{
This package provides predictive microbiology models to allow users to predict the behavior of pathogenic bacteria in foods.
}
\author{
\strong{Maintainer}: Ghaya Ben Hmidene \email{ghaya.benhmidene@gmail.com}

Authors:
\itemize{
  \item Laurent Guillier \email{guillier.laurent@gmail.com} [contributor]
}

}
\keyword{internal}
