% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranyi_diff_without_lag.R
\name{baranyi_diff_without_lag}
\alias{baranyi_diff_without_lag}
\title{The differential form of the predictive growth model developed by Baranyi & Roberts in 1994, to predict microbial growth in food environments.}
\usage{
baranyi_diff_without_lag(t, state, parameters)
}
\arguments{
\item{t}{time vector}

\item{state}{initial values of y  for the first time value.}

\item{parameters}{parameter list}
}
\value{
dy
}
\description{
The differential form of the predictive growth model developed by Baranyi & Roberts in 1994, to predict microbial growth in food environments.
}
\details{
\deqn{\begin{equation}d y=\operatorname{mu_{max}}^{*}(1-(y / y_{max} ))^{*} y\end{equation}}
}
\examples{
library (deSolve)
state<-c(y=15)
time<-seq(0,70, 1)
parameters<- c(mumax=1, ymax=10000)
out_baranyi_without_lag<-ode(y = state, times = time, func = baranyi_diff_without_lag,
 parms = parameters)
plot(time,log10(out_baranyi_without_lag[,"y"]),lty=1,col='red')
}
