% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranyi_diff_without_Nmax.R
\name{baranyi_diff_without_Nmax}
\alias{baranyi_diff_without_Nmax}
\title{The differential form of the predictive growth model developed by Baranyi & Roberts in 1994, to predict microbial growth in food environments.}
\usage{
baranyi_diff_without_Nmax(t, state, parameters)
}
\arguments{
\item{t}{time vector}

\item{state}{initial values of y and Q for the first time value.}

\item{parameters}{parameter list}
}
\value{
dQ, dy
}
\description{
The differential form of the predictive growth model developed by Baranyi & Roberts in 1994, to predict microbial growth in food environments.
}
\details{
\deqn{\begin{equation}\mathrm{d} \mathrm{Q}=\mathrm{mu_{max}}^{*} \mathrm{Q}\end{equation}}
\deqn{\begin{equation}d y=(Q /(1+Q))^{*} \operatorname{mu_{max}}^{*} y\end{equation}}
}
\examples{
library(deSolve)
K=9.21
Q=1/((1/exp(-K))-1)
state<-c(Q=Q,y=15)
time<-seq(0,70, 1)
parameters<- c(mumax=1)
out_baranyi_without_lag<-ode(y = state, times = time, func = baranyi_diff_without_Nmax, parms = parameters)
plot(time,log10(out_baranyi_without_lag[,"y"]),lty=1,col='red')
}
