% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapH.R
\docType{data}
\name{datapH}
\alias{datapH}
\title{datapH}
\format{
A data frame with 13 rows and 2 variables.
}
\source{
\url{https://sfamjournals.onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2672.2010.04858.x}
}
\usage{
data("datapH")
}
\description{
A dataset containing 13 observations of growth rate of Listeria innocua for 13 pH values
}
\details{
this dataset was extracted from the review of Ronald J.W. Lambert (2010) entitled: A new model for the effect of pH on microbial growth: an extension of the Gamma hypothesis
}
\references{
https://sfamjournals.onlinelibrary.wiley.com/doi/epdf/10.1111/j.1365-2672.2010.04858.x
}
\keyword{datasets}
