% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_aw_2p.R
\name{Gamma_aw_2p}
\alias{Gamma_aw_2p}
\title{this function of secondary growth model describe the evolution of the square root of the maximum specific growth rate (sqrtmumax) as a function of water activity (aw), This is a symetric cardinal pH model developped by Rosso & Robinson in 2001 with two parameters (awmin, muopt) and awopt =1}
\usage{
Gamma_aw_2p(aw, awmin, muopt)
}
\arguments{
\item{aw}{water activity #a number}

\item{awmin}{minimal growth water activity # a number}

\item{muopt}{Optimal growth rate # a number # a number}
}
\value{
sqrmumax^2= mumax #maximum growth rate # a number
}
\description{
this function of secondary growth model describe the evolution of the square root of the maximum specific growth rate (sqrtmumax) as a function of water activity (aw), This is a symetric cardinal pH model developped by Rosso & Robinson in 2001 with two parameters (awmin, muopt) and awopt =1
}
\examples{
Gamma_aw_2p(0.92,0.90,0.097)
}
