% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_aw_3p.R
\name{Gamma_aw_3p}
\alias{Gamma_aw_3p}
\title{this function of secondary growth model describe the evolution of the square root of the maximum specific growth rate (sqrtmumax) as a function of water activity (aw), This is a symetric cardinal pH model developped by Rosso & Robinson in 2001 with three parameters (awmin, awopt, muopt)}
\usage{
Gamma_aw_3p(aw, awmin, awopt, muopt)
}
\arguments{
\item{aw}{water activity #a number}

\item{awmin}{minimal growth water activity # a number}

\item{awopt}{optimal growth water activity # a number}

\item{muopt}{Optimal growth rate # a number # a number}
}
\value{
sqrmumax^2= mumax #maximum growth rate # a number
}
\description{
this function of secondary growth model describe the evolution of the square root of the maximum specific growth rate (sqrtmumax) as a function of water activity (aw), This is a symetric cardinal pH model developped by Rosso & Robinson in 2001 with three parameters (awmin, awopt, muopt)
}
\examples{
Gamma_aw_3p(0.95,0.90, 0.98, 0.5)
}
