% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_pH_3p.R
\name{Gamma_pH_3p}
\alias{Gamma_pH_3p}
\title{this function of secondary growth model describe the evolution of the square root of the maximum specific growth rate (sqrtmumax) as a function of pH, This is a symetric cardinal pH model developped by Rosso & al.in 1995  with three parameters (pHmin, pHopt, muopt), obtained by fixing pHmax =2}
\usage{
Gamma_pH_3p(pH, pHmin, muopt, pHopt)
}
\arguments{
\item{pH}{# a number}

\item{pHmin}{Minimal growth pH #a number}

\item{muopt}{Optimal growth rate # a number}

\item{pHopt}{Optimal growth pH #a number}
}
\value{
sqrmumax^2= mumax #maximum growth rate # a number
}
\description{
this function of secondary growth model describe the evolution of the square root of the maximum specific growth rate (sqrtmumax) as a function of pH, This is a symetric cardinal pH model developped by Rosso & al.in 1995  with three parameters (pHmin, pHopt, muopt), obtained by fixing pHmax =2
}
\examples{
Gamma_pH_3p(5,4,0.5,7)
Gamma_pH_3p(4,4,0.2,7)
Gamma_pH_3p(7,4,2,7)
}
