% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBaran.R
\name{modBaran}
\alias{modBaran}
\title{The primary predictive model developped by Baranyi & Roberts in 1994, to predict microbial behavior in food environments.}
\usage{
modBaran(temp, mumax, lag, log10N0, log10Nmax)
}
\arguments{
\item{temp}{time(h)}

\item{mumax}{maximum growth rate # a number}

\item{lag}{lag time # a number}

\item{log10N0}{Initial population density #a number}

\item{log10Nmax}{maximum population density # a number}
}
\value{
log10N #a number
}
\description{
The primary predictive model developped by Baranyi & Roberts in 1994, to predict microbial behavior in food environments.
}
\examples{
modBaran(temp=seq(0,24,0.5), mumax = 0.037, lag=4, log10N0 = 1, log10Nmax = 5 )
}
