% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBaran_without_lag.R
\name{modBaran_without_lag}
\alias{modBaran_without_lag}
\title{This is the primary predictive model developped by Baranyi & Roberts in 1994, to predict microbial behavior in food environments.The model modBaran_without_lag has four parameters (temps, mumax, log10N0, log10Nmax) without lag phase.}
\usage{
modBaran_without_lag(temp, mumax, log10N0, log10Nmax)
}
\arguments{
\item{temp}{time(h)}

\item{mumax}{maximum growth rate # a number}

\item{log10N0}{Initial population density #a number}

\item{log10Nmax}{maximum population density # a number}
}
\value{
log10N #a number
}
\description{
This is the primary predictive model developped by Baranyi & Roberts in 1994, to predict microbial behavior in food environments.The model modBaran_without_lag has four parameters (temps, mumax, log10N0, log10Nmax) without lag phase.
}
\details{
\deqn{\begin{equation}\begin{array}{l}
\mathrm{LOG} 10 \mathrm{N}=\left(\mathrm{LOG} 10 \mathrm{Nmax}-\log 10\left(1+\left(10_{\mathrm{m}}^{\wedge}(\mathrm{LOG} 10 \mathrm{Nmax}-\mathrm{LOG} 10 \mathrm{N} 0)-1\right)\right.\right. \\
\left.\left.* \exp \left(-\operatorname{mumax}^{*} \mathrm{t}\right)\right)\right)
\end{array}\end{equation}}
}
\examples{
modBaran_without_lag(seq(0,24,1), 0.5,3,5)
}
