% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBaran_without_Nmax.R
\name{modBaran_without_Nmax}
\alias{modBaran_without_Nmax}
\title{This is the primary predictive model developped by Baranyi & Roberts in 1994, to predict microbial behavior in food environments.The model modBaran_without_lag has four parameters (temps, mumax, log10N0, log10Nmax) without Nmax}
\usage{
modBaran_without_Nmax(temp, mumax, lag, log10N0)
}
\arguments{
\item{temp}{time (h)}

\item{mumax}{maximum growth rate # a number}

\item{lag}{lag time # a number}

\item{log10N0}{Initial population density #a number}
}
\value{
log10N #a number
}
\description{
This is the primary predictive model developped by Baranyi & Roberts in 1994, to predict microbial behavior in food environments.The model modBaran_without_lag has four parameters (temps, mumax, log10N0, log10Nmax) without Nmax
}
\details{
\deqn{\begin{equation}\begin{array}{l}
\mathrm{LOG} 10 \mathrm{N}=\mathrm{LOG} 10 \mathrm{N} 0+\operatorname{mumax}^{*} \mathrm{t} / \log (10)+\log 10\left(\exp \left(-\operatorname{mumax}^{*} \mathrm{t}\right) *\right. \\
\left.\left(1-\exp \left(-\operatorname{mumax}^{*} \mathrm{Lag}\right)\right)+\exp \left(-\operatorname{mumax}^{*} \mathrm{Lag}\right)\right)
\end{array}\end{equation}}
}
\examples{
 modBaran_without_Nmax(seq(0,48,0.5), 0.035, 4, 3)
}
