% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBuchanan.R
\name{modBuchanan}
\alias{modBuchanan}
\title{The primary predictive model developped by Buchanan in 1918, to predict microbial behavior in food environments.}
\usage{
modBuchanan(temp, log10N0, log10Nmax, mumax, lag)
}
\arguments{
\item{temp}{time vector}

\item{log10N0}{Initial population density #a number}

\item{log10Nmax}{maximum population density # a number}

\item{mumax}{maximum growth rate # a number}

\item{lag}{lag time # a number}
}
\value{
LOG10N
}
\description{
The primary predictive model developped by Buchanan in 1918, to predict microbial behavior in food environments.
}
\details{
\deqn{\begin{equation}\operatorname{Ln}[N(t)]=\ln \left(N_{0}\right) \quad \text { si } t \leq \lambda\end{equation}}
\deqn{\begin{equation}
\operatorname{Ln}[N(t)]=\ln \left(N_{0}\right)-\mu(t-\lambda) \quad \text { si } t>\lambda
\end{e\begin{equation}
\operatorname{Ln}[N(t)]=\ln \left(N_{\max }\right) \quad \operatorname{si} N(t) \geq N_{\max }
\end{equation}quation}}
\deqn{}
}
\examples{
temp<-seq(0,48,0.5)
modBuchanan(temp,1,9,0.5,4)
}
