% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modELPM.R
\name{modELPM}
\alias{modELPM}
\title{This function describes the extended Lambert–Pearson model (ELPM), who model the effect of pH (as hydrogen and hydroxyl ions) over the whole biokinetic pH range.}
\usage{
modELPM(pH, P1, P2, P3, P4, muopt)
}
\arguments{
\item{pH}{pH #a number}

\item{P1}{a parameter of the ELPM model # a number}

\item{P2}{a parameter of the ELPM model # a number}

\item{P3}{a parameter of the ELPM model # a number}

\item{P4}{a parameter of the ELPM model # a number}

\item{muopt}{Optimal growth rate # a number a number}
}
\value{
mumax #maximum growth rate # a number
}
\description{
This function describes the extended Lambert–Pearson model (ELPM), who model the effect of pH (as hydrogen and hydroxyl ions) over the whole biokinetic pH range.
}
\details{
\deqn{\begin{equation}\begin{aligned}
\frac{\mathrm{d} \gamma}{\mathrm{dpH}}=& 2.303\left\{P_{2}\left(\frac{10^{-\mathrm{pH}}}{P_{1}}\right)^{P_{2}}-P_{4}\left(\frac{10^{\mathrm{pH}-14}}{P_{3}}\right)^{P_{4}}\right\} \\
& \exp \left(-\left(\frac{10^{-\mathrm{pH}}}{P_{1}}\right)^{P_{2}}+\left(\frac{10^{\mathrm{pH}-14}}{P_{3}}\right)^{P_{4}}\right)
\end{aligned}\end{equation}}
}
\examples{
modELPM( 7.585261, 0.000132, 0.889, 0.0004, 0.0000005, 0.5)
}
