% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPresserext.R
\name{modelPresserext}
\alias{modelPresserext}
\title{This extended Presser model was quoted by Tienungoon et al. (2000) for the full biokinetic range}
\usage{
modelPresserext(pH, pHmin, pHmax, muopt)
}
\arguments{
\item{pH}{pH # a number}

\item{pHmin}{Minimal growth pH #a number}

\item{pHmax}{Maximal growth pH #a number}

\item{muopt}{Optimal growth rate # a number}
}
\value{
mumax #maximum growth rate # a number
}
\description{
This extended Presser model was quoted by Tienungoon et al. (2000) for the full biokinetic range
}
\details{
\deqn{\begin{equation}\frac{\mu_{\max }}{\mu_{\mathrm{opt}}}=\gamma_{\mathrm{pH}}=\left(1-10^{\mathrm{pH}_{\min }-\mathrm{pH}}\right)\left(1-10^{\mathrm{pH}-\mathrm{pH}_{\max }}\right)\end{equation}}
}
\examples{
modelPresserext(7,3,9,0.5)
modelPresserext(c(7,5),3,9,0.5)
}
