% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modrosso.R
\name{modrosso}
\alias{modrosso}
\title{The logistic primary growth model using delay and rupture, developped by Rosso in 1996, describes growth as a function of time.}
\usage{
modrosso(temp, mumax, lag, Nmax, N0)
}
\arguments{
\item{temp}{time (h)}

\item{mumax}{maximum growth rate # a number}

\item{lag}{lag time # a number}

\item{Nmax}{maximum population size #a number}

\item{N0}{initial population size # a number}
}
\value{
log10N #a number
}
\description{
The logistic primary growth model using delay and rupture, developped by Rosso in 1996, describes growth as a function of time.
}
\examples{
modrosso(temp=seq(0,24,0.5),mumax=0.37,lag=4,Nmax = 10^5,N0=100)
}
