% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modZwite.R
\name{modZwite}
\alias{modZwite}
\title{This model is developed by  Zwietering et al. (1992, 1993).It's a square root type model but with an extra ﬁtting parameter-c2}
\usage{
modZwite(pH, pHmin, pHmax, pHopt, muopt, c2)
}
\arguments{
\item{pH}{pH # a number}

\item{pHmin}{Minimal growth pH #a number}

\item{pHmax}{Maximal growth pH #a number}

\item{pHopt}{Optimal growth pH #a number}

\item{muopt}{Optimal growth rate # a number}

\item{c2}{a parameter # a number}
}
\value{
mumax #maximum growth rate # a number
}
\description{
This model is developed by  Zwietering et al. (1992, 1993).It's a square root type model but with an extra ﬁtting parameter-c2
}
\details{
\deqn{\begin{equation}\gamma_{\mathrm{pH}}=\left[\frac{\left(\mathrm{pH}-\mathrm{pH}_{\min }\right)\left(1-\exp \left\{\mathrm{c}_{2}\left(\mathrm{pH}-\mathrm{pH}_{\max }\right)\right\}\right)}{\left(\mathrm{pH}_{\mathrm{opt}}-\mathrm{pH}_{\min }\right)\left(1-\exp \left\{\mathrm{c}_{2}\left(\mathrm{pH}_{\mathrm{opt}}-\mathrm{pH}_{\max }\right)\right\}\right)}\right]^{2}\end{equation}}
}
\examples{
modZwite(5,3,9,7,0.5,0.4)
modZwite(c(5,7),3,9,7,0.5,0.4)
}
