% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proba_aw.R
\name{Proba_aw}
\alias{Proba_aw}
\title{This function describes the evolution of the single-cell growth probability with water activity (aw). For this, we used the model proposed by Jean Christophe Augustin and Aurelia Czarnecka-Kwasiborski in 2012, to describe the increase of the probability from 0 to 1 with increasing values of aw.}
\usage{
Proba_aw(aw, awinf, awsup)
}
\arguments{
\item{aw}{water activity value #a number}

\item{awinf}{The cellular minimal value of aw for growth # a number}

\item{awsup}{The cellular maximal value of aw for growth # a number}
}
\value{
proba_pH_Aug The single-cell growth probability # a number
}
\description{
This function describes the evolution of the single-cell growth probability with water activity (aw). For this, we used the model proposed by Jean Christophe Augustin and Aurelia Czarnecka-Kwasiborski in 2012, to describe the increase of the probability from 0 to 1 with increasing values of aw.
}
