% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proba_c.R
\name{Proba_c}
\alias{Proba_c}
\title{This function describes the evolution of the probability of single-cell growth with environmental factors (T, pH & aw). For this, we used the model proposed by Jean Christophe Augustin and Aurelia Czarnecka-Kwasiborski in 2012, to describe the combined effect of the three environmental factors. in this model T, pH and aw were considered to have independent effects on the single cell}
\usage{
Proba_c(T, Tinf, Tsup, c, pH, pHinf, pHsup, aw, awinf, awsup)
}
\arguments{
\item{T}{Temperature value # a number}

\item{Tinf}{The theoritical cellular minimal temperature for growth #a number}

\item{Tsup}{The theoritical cellular maximal temperature for growth #a number}

\item{c}{Parameter # a number}

\item{pH}{pH value # a number}

\item{pHinf}{The cellular minimal pH for growth # a number}

\item{pHsup}{The cellular maximal pH for growth # a number}

\item{aw}{water activity value #a number}

\item{awinf}{The cellular minimal value of aw for growth # a number}

\item{awsup}{The cellular maximal value of aw for growth # a number}
}
\value{
Pc The single-cell growth probability as function of environemental factors # a number
}
\description{
This function describes the evolution of the probability of single-cell growth with environmental factors (T, pH & aw). For this, we used the model proposed by Jean Christophe Augustin and Aurelia Czarnecka-Kwasiborski in 2012, to describe the combined effect of the three environmental factors. in this model T, pH and aw were considered to have independent effects on the single cell
}
\examples{
Proba_c(10,Tinf=-3.6,Tsup=17.3,5,c=7.6,pHinf=-4.34,pHsup=5.93,0.95,awinf=0.917,awsup=0.988)
}
