% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proba_pH.R
\name{Proba_pH}
\alias{Proba_pH}
\title{This function describes the evolution of the single-cell growth probability with pH. For this, we used the model proposed by Jean Christophe Augustin and Aurelia Czarnecka-Kwasiborski in 2012, to describe the increase of the probability from 0 to 1 with increasing values of pH.}
\usage{
Proba_pH(pH, pHinf, pHsup)
}
\arguments{
\item{pH}{pH value # a number}

\item{pHinf}{The cellular minimal pH for growth # a number}

\item{pHsup}{The cellular maximal pH for growth # a number}
}
\value{
proba_pH_Aug The single-cell growth probability # a number
}
\description{
This function describes the evolution of the single-cell growth probability with pH. For this, we used the model proposed by Jean Christophe Augustin and Aurelia Czarnecka-Kwasiborski in 2012, to describe the increase of the probability from 0 to 1 with increasing values of pH.
}
