% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_T.R
\name{Gamma_T}
\alias{Gamma_T}
\title{This secondary growth model describes the influence of temperature over mumax}
\usage{
Gamma_T(T, Tmin, Topt, Tmax, muopt)
}
\arguments{
\item{T}{Temperature # a number}

\item{Tmin}{Minimal growth temperature #a number}

\item{Topt}{Optimal growth temperature #a number}

\item{Tmax}{Maximal growth temperature # a number}

\item{muopt}{optimal growth rate # a number}
}
\value{
sqrmumax^2= mumax #maximum growth rate # a number
}
\description{
This secondary growth model describes the influence of temperature over mumax
}
\examples{
Gamma_T(10,4,20,37,0.037)
}
