% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modrosso.R
\name{modrosso}
\alias{modrosso}
\title{The logistic primary growth model using delay and rupture, developped by Rosso in 1996, describes growth as a function of time.}
\usage{
modrosso(temp, mumax, lag, N0, Nmax)
}
\arguments{
\item{temp}{time (h)}

\item{mumax}{maximum growth rate # a number}

\item{lag}{lag time # a number}

\item{N0}{initial population size # a number}

\item{Nmax}{maximum population size #a number}
}
\value{
log10N #a number
}
\description{
The logistic primary growth model using delay and rupture, developped by Rosso in 1996, describes growth as a function of time.
}
\details{
\deqn{\begin{equation}f\left(t, \Theta_{1}\right)=\left\{\begin{array}{ll}
\ln x_{0} & , t \leq \operatorname{lag} \\
\ln x_{\max }-\ln \left(1+\left(\frac{x_{\max }}{x_{0}}-1\right) \cdot \exp \left(-\mu_{\max } \cdot(t-\operatorname{lag})\right)\right) & , t>\operatorname{lag}
\end{array}\right.\end{equation}}
}
\examples{
modrosso(temp=seq(0,24,0.5),mumax=0.37,lag=4,N0=100,Nmax = 10^5)
}
